RAC-721: Move Public API in component/query and implementation in bundle
