+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/queries.yml
@@ -1,12 +1,12 @@
 services:
     akeneo.enrichment.public_api.find_category_trees:
-        class: Akeneo\Pim\Enrichment\PublicApi\Categories\SqlFindCategoryTrees
+        class: Akeneo\Pim\Enrichment\Bundle\Storage\Sql\Category\SqlFindCategoryTrees
         arguments:
             - '@pim_catalog.repository.category'
             - '@pim_catalog.normalizer.standard.translation'
 
     akeneo.enrichment.public_api.count_categories_per_tree:
-        class: Akeneo\Pim\Enrichment\PublicApi\Categories\SqlCountCategoriesPerTree
+        class: Akeneo\Pim\Enrichment\Bundle\Storage\Sql\Category\SqlCountCategoriesPerTree
         arguments:
             - '@database_connection'
