+++ src/Akeneo/Pim/Enrichment/Component/Category/Query/PublicApi/CategoryTree.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Pim\Enrichment\PublicApi\Categories;
+namespace Akeneo\Pim\Enrichment\Component\Category\Query\PublicApi;
 
 /**
  * @author    Samir Boulil <samir.boulil@akeneo.com>
